Fix imports for server to not pull in client only code
