+++ src/dialogs/helpers.ts
@@ -24,7 +24,7 @@ class Debouncer {
     if (this.timeout) {
       clearTimeout(this.timeout);
     }
-    this.timeout = setTimeout(func, this.delay);
+    this.timeout = setTimeout(func, this.delay) as any;
   }
 }
 
@@ -47,4 +47,17 @@ class EagerDebouncer {
   }
 }
 
-export { disableTooExpensive, Debouncer, EagerDebouncer };
+const sideBySideDivs = (content: string[], shrink = false) => {
+  let ret = `<div style="display: flex; flex-direction: row; justify-content: left;"> ${content
+    .map((c) => `<div style="flex: 1;">${c}</div>`)
+    .join("")}</div>`;
+  if (shrink) {
+    return `<div style="display: inline-block;">${ret}</div>`;
+  }
+  return ret;
+};
+
+const stackedDivs = (content: string[]) => `<div style="display: flex; flex-direction: column; justify-content: left;">
+  ${content.map((c) => `<div style="flex: 1;">${c}</div>`).join("")}</div>`;
+
+export { disableTooExpensive, Debouncer, EagerDebouncer, sideBySideDivs, stackedDivs };
